// packages/datadog-api-client-v1/models/SyntheticsAPITestConfig.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SyntheticsAPIStep } from './SyntheticsAPIStep';
import { SyntheticsAssertion } from './SyntheticsAssertion';
import { SyntheticsConfigVariable } from './SyntheticsConfigVariable';
import { SyntheticsTestRequest } from './SyntheticsTestRequest';
/**
* Configuration object for a Synthetic API test.
*/
export declare class SyntheticsAPITestConfig {
    /**
    * Array of assertions used for the test.
    */
    'assertions': Array<SyntheticsAssertion>;
    /**
    * Array of variables used for the test.
    */
    'configVariables'?: Array<SyntheticsConfigVariable>;
    'request'?: SyntheticsTestRequest;
    /**
    * When the test subtype is `multi`, the steps of the test.
    */
    'steps'?: Array<SyntheticsAPIStep>;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsAPITestConfig;
    static serialize(data: SyntheticsAPITestConfig): {
        [key: string]: any;
    };
    constructor();
}